PG ci osx

+++ scripts/osx/ci_build_pg.sh
@@ -1,15 +1,15 @@
 #!/bin/bash
-set -e
+# set -e
 
 CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-SCRIPT_DIR="$( cd "$( dirname "${CURRENT_DIR}"/../../ )" && pwd )"
-PG_DIR="$( cd "$( dirname "${SCRIPT_DIR}/../../" )" && pwd )"
+export SCRIPT_DIR="$( cd "$( dirname "${CURRENT_DIR}"/../../ )" && pwd )"
+export PG_DIR="$( cd "$( dirname "${SCRIPT_DIR}/../../" )" && pwd )"
 
-OF_DIR="$( cd "$( dirname "${PG_DIR}/../../../" )" && pwd )"
+export OF_DIR="$( cd "$( dirname "${PG_DIR}/../../../" )" && pwd )"
 
-FRONTEND_DIR="$( cd "$( dirname "${PG_DIR}/frontend" )" && pwd )"
+export FRONTEND_DIR="$( cd "$( dirname "${PG_DIR}/frontend" )" && pwd )"
 
-CMDLINE_DIR="$( cd "$( dirname "${PG_DIR}/commandLine" )" && pwd )"
+export CMDLINE_DIR="$( cd "$( dirname "${PG_DIR}/commandLine" )" && pwd )"
 
 
 echo "CURRENT_DIR:  ${CURRENT_DIR}"
@@ -36,25 +36,24 @@ echoDots(){
 }
 
 package_app(){
-	if [[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" ]] ; then
-		echo "package_app -- $POSTFIX"
+	# if [[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" ]] ; then
+		echo "package_app -- [${PLATFORM}${POSTFIX}]"
 		PLATFORM=$1
 		# Copy commandLine into electron .app
-		cd ${PG_DIR}
-		echo "Current directory: (should be projectGenerator)"
+		cd "${PG_DIR}"
+		echo "Current directory: (should be projectGenerator) [${PG_DIR}]"
 		pwd
 		echo "Directory contents:"
 		ls
 		echo "-------------"
 		echo "copy cmdline PG from app contents (signed)"
 		if command -v rsync &> /dev/null; then
-			rsync -avzp "commandLine/bin/commandLine.app/Contents/MacOS/commandLine" "projectGenerator-$PLATFORM$POSTFIX/projectGenerator.app/Contents/Resources/app/app/projectGenerator" 2> /dev/null
+			rsync -avzp "${PG_DIR}/commandLine/bin/commandLine.app/Contents/MacOS/commandLine" "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app/Contents/Resources/app/app/projectGenerator" 2> /dev/null
 		else
-			cp -aX commandLine/bin/commandLine.app/Contents/MacOS/commandLine projectGenerator-$PLATFORM$POSTFIX/projectGenerator.app/Contents/Resources/app/app/projectGenerator 2> /dev/null
+			cp -aX "${PG_DIR}/commandLine/bin/commandLine.app/Contents/MacOS/commandLine" "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app/Contents/Resources/app/app/projectGenerator" 2> /dev/null
 		fi
 
-
-		cd ${PG_DIR}
+		cd "${PG_DIR}"
 		pwd
 		echo "Directory contents:"
 		ls
@@ -62,24 +61,33 @@ package_app(){
 		# sed -i -e "s/osx/$PLATFORM/g" projectGenerator-$PLATFORM/projectGenerator.app/Contents/Resources/app/settings.json
 
 		# Sign app
-		echo "Signing electron .app"
-		echo "cd to ${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
-		cd ${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}
+		echo "[Signing electron] .app"
+		echo "cd to [${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}]"
+		cd "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
 		pwd
 		echo "Directory contents:"
 		ls
 		echo "-------------"
 		xattr -cr projectGenerator.app
 		echo "-------------"
-		echo "cd to ${PG_DIR}"
-		cd ${PG_DIR}
+		echo "cd to [${PG_DIR}]"
+		cd "${PG_DIR}"
+		pwd
 
-		cd projectGenerator.app/Contents/MacOS/projectGenerator
-		${SCRIPT_DIR}/secure.sh projectGenerator
-		cd ${PG_DIR}
-		cd projectGenerator.app/Contents/Resources/app/app/projectGenerator
-		${SCRIPT_DIR}/secure.sh projectGenerator
-		cd ../
+		echo "[Secure] creating PKL files"
+		cd "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app/Contents/MacOS/projectGenerator"
+		pwd
+
+
+		. "${SCRIPT_DIR}/secure.sh"
+    	secure "projectGenerator" projectGenerator.pkl
+
+		cd "${PG_DIR}"
+		cd "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app/Contents/Resources/app/app/projectGenerator"
+		pwd
+		. "${SCRIPT_DIR}/secure.sh"
+    	secure "projectGenerator" projectGenerator.pkl
+		cd "${PG_DIR}"
 
 		TEAM_ID="HC25N2E7UT"
 		APPLE_ID="theo@theowatson.com"
@@ -87,26 +95,34 @@ package_app(){
 		# echo "--identity=3rd Party Mac Developer Application: ${APPLE_ID} (${TEAM_ID})"
 		if [[ ("${TRAVIS_REPO_SLUG}/${TRAVIS_BRANCH}" == "openframeworks/projectGenerator/master" || "${TRAVIS_REPO_SLUG}/${TRAVIS_BRANCH}" == "openframeworks/projectGenerator/bleeding") && "$TRAVIS_PULL_REQUEST" == "false" ]] || 
    			[[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" ]] ; then
-			electron-osx-sign projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app --platform=darwin --type=distribution --no-gatekeeper-assess --hardened-runtime --entitlements=scripts/osx/PG.entitlements --entitlements-inherit=scripts/osx/PG.entitlements
+			electron-osx-sign "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app" --platform=darwin --type=distribution --no-gatekeeper-assess --hardened-runtime --entitlements=${SCRIPT_DIR}/osx/PG.entitlements --entitlements-inherit=${SCRIPT_DIR}/osx/PG.entitlements
 		fi
 		
-		echo "Compressing PG app"
+		echo "[Compressing] projectGenerator gui"
 		# need to upload zip of just app to apple for notarizing
-		zip --symlinks -r -q projectGenerator-$PLATFORM$POSTFIX/projectGenerator-$PLATFORM$POSTFIX.zip projectGenerator-$PLATFORM$POSTFIX/projectGenerator.app
+		zip --symlinks -r -q "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator-${PLATFORM}${POSTFIX}.zip" "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator.app"
 		if [[ ("${TRAVIS_REPO_SLUG}/${TRAVIS_BRANCH}" == "openframeworks/projectGenerator/master" || "${TRAVIS_REPO_SLUG}/${TRAVIS_BRANCH}" == "openframeworks/projectGenerator/bleeding") && "$TRAVIS_PULL_REQUEST" == "false" ]] || 
    			[[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" ]] ; then
-			xcrun notarytool submit "projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator-${PLATFORM}${POSTFIX}.zip" --apple-id "${APPLE_ID}" --team-id "${TEAM_ID}" --password "${GA_APPLE_PASS}"
+			echo "[Notorizing] projectGenerator gui [projectGenerator-${PLATFORM}${POSTFIX}]"
+			xcrun notarytool submit "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator-${PLATFORM}${POSTFIX}.zip" --apple-id "${APPLE_ID}" --team-id "${TEAM_ID}" --password "${GA_APPLE_PASS}"
 		fi
-
-		mv projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator-${PLATFORM}${POSTFIX}.zip ${PG_DIR}/../../../projectGenerator/projectGenerator-${PLATFORM}${POSTFIX}.zip
-		cd ${PG_DIR}/../../../projectGenerator
-		echo "Final Directory contents: ${PG_DIR}/../../../projectGenerator"
+		cd "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
 		pwd
 		ls
-		echo "-------------"
-	else
-		echo "package_app not on master/bleeding so will not package app"	
-	fi
+
+		cd "${PG_DIR}"
+		if [[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" ]] ; then
+		
+			mv -v "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}/projectGenerator-${PLATFORM}${POSTFIX}.zip" "${PG_DIR}/../../../projectGenerator/projectGenerator-${PLATFORM}${POSTFIX}.zip"
+			cd ${PG_DIR}/../../../projectGenerator
+			echo "Final Directory contents: [${PG_DIR}/../../../projectGenerator]"
+			pwd
+			ls
+			echo "-------------"
+		fi
+	# else
+	# 	echo "package_app not on master/bleeding so will not package app"	
+	# fi
 }
 
 
@@ -164,7 +180,7 @@ sign_and_upload(){
 
 import_certificate(){
 
-	echo "  import_certificate"
+	echo "  [import_certificate]"
 
 	if [[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" && -n "${CERTIFICATE_OSX_APPLICATION}" ]]; then
 		echo "Decoding signing certificates"
@@ -192,7 +208,7 @@ import_certificate(){
 		# makes keychain not timeout
 		security set-keychain-settings $KEY_CHAIN
 
-		echo "import certificates done"
+		echo "[import certificates done]"
 	fi
 
 }
@@ -202,56 +218,83 @@ echo "##[group]build_cmdline.sh"
 ${CURRENT_DIR}/build_cmdline.sh
 echo "##[endgroup]"
 
-
 echo "##[group]import_certificate"
 import_certificate
-# Generate electron app
 echo "##[endgroup]"
 
-echo "##[group]build_frontend"
+# Generate electron app
+# echo "##[group]build_frontend arm64"
+# ${CURRENT_DIR}/build_frontend_arm64.sh
+# echo "##[endgroup]"
+
+# echo "##[group]build_frontend x86_64"
+# ${CURRENT_DIR}/build_frontend_64.sh
+# echo "##[endgroup]"
+
+echo "##[group]build_frontend mac-universal"
 ${CURRENT_DIR}/build_frontend.sh
 echo "##[endgroup]"
-
-echo "Current directory: (should be projectGenerator)"
-pwd
-echo "Directory contents:"
-ls
 
 if [ -d "${PG_DIR}/projectGenerator-osx" ]; then
 	rm -rf ${PG_DIR}/projectGenerator-osx
 fi
 
-cd ${PG_DIR}
-echo "Current directory: (should be apps/projectGenerator)"
+if [ -d "${PG_DIR}/projectGenerator-osx_arm64" ]; then
+	rm -rf ${PG_DIR}/projectGenerator-osx_arm64
+fi
+
+if [ -d "${PG_DIR}/projectGenerator-osx_64" ]; then
+	rm -rf ${PG_DIR}/projectGenerator-osx_64
+fi
+
+cd "${PG_DIR}"
+echo "Current directory: (should be apps/projectGenerator) after builds [ ${PG_DIR} ]"
 pwd
 echo "Directory contents:"
 ls
 echo "-------------"
-if [ -d "${MACOS_ARM64}" ]; then
-	POSTFIX="_arm64"
-	mv frontend/dist/mac-arm64 ${PG_DIR}/projectGenerator-osx$POSTFIX
-else
-	POSTFIX=""
-	mv frontend/dist/mac-universal ${PG_DIR}/projectGenerator-osx
+
+PLATFORM=osx
+
+# export POSTFIX="_arm64"
+# if [ -z "${PG_DIR}/projectGenerator-osx${POSTFIX}" ]; then
+# 	mkdir -v "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
+# 	mv -v "${PG_DIR}/frontend/dist/mac-arm64" "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
+# fi
+# echo "-------------"
+# echo "  package_app osx arm64"
+# package_app ${PLATFORM}
+
+
+# echo "-------------"
+# export POSTFIX="_64"
+# if [ -z "${PG_DIR}/projectGenerator-osx${POSTFIX}" ]; then
+# 	mkdir -v "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
+# 	mv -v "${PG_DIR}/frontend/dist/mac" "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
+# fi
+# echo "-------------"
+# echo "  package_app osx 64"
+# package_app ${PLATFORM}
+
+
+export POSTFIX=""
+if [ -z "${PG_DIR}/projectGenerator-osx${POSTFIX}" ]; then
+	mkdir -v "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
+	mv -v "${PG_DIR}/frontend/dist/mac-universal" "${PG_DIR}/projectGenerator-${PLATFORM}${POSTFIX}"
+fi
+echo "-------------"
+echo "  package_app osx universal"
+package_app ${PLATFORM}
+echo "-------------"
+
+if [[ ("${GITHUB_REF##*/}" == "master" || "${GITHUB_REF##*/}" == "bleeding") && -z "${GITHUB_HEAD_REF}" && -n "${CERTIFICATE_OSX_APPLICATION}" ]]; then
+		
+	echo "  [security] remove imported keys."
+	rm -rf scripts/id_rsa 2> /dev/null
+	rm -rf scripts/*.p12 2> /dev/null
+
 fi
 
 
-
-echo "-------------"
-echo "  package_app osx"
-package_app osx
-
-# echo "package_app ios"
-# npm run build:macos > /dev/null
-
-# if [ -d "${PG_DIR}/projectGenerator-ios" ]; then
-# 	rm -rf ${PG_DIR}/projectGenerator-ios
-# fi
-# mv dist/mac ${PG_DIR}/projectGenerator-ios
-# echo "package_app ios"
-# package_app ios
-
-echo "  for security remove keys maybe imported"
-rm -rf scripts/id_rsa 2> /dev/null
-rm -rf scripts/*.p12 2> /dev/null
+echo "  [Completed]"
